backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlmodel==0.0.22
psycopg[binary]==3.2.13
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
pydantic-settings==2.4.0
openai

groq
youtube-transcript-api
pypdf
python-multipart